Clean up config components (keys.hpp and led.hpp)
