--- include/config/led_configuration.hpp
@@ -1,17 +1,0 @@
-#pragma once
-
-#include "config/components/led_effect_type.hpp"
-#include "config/components/led.hpp"
-
-// Configuration for the LEDs, containing the actual LEDs, brightness, effects, ... of the keypad.
-struct LedConfiguration
-{
-    // A list of all LEDs. (rgb, effect, ...)
-    Led leds[LEDS];
-
-    // The brightness of the LEDs. (0-100)
-    uint8_t brightness;
-
-    // The ID of the RGB effect.
-    LedEffectType effect;
-};
